feat(catalogue): improve by reference query logical,add delete by reference api

+++ src/Hx.Abp.Attachment.EntityFrameworkCore/Hx/Abp/Attachment/EntityFrameworkCore/EfCoreAttachCatalogueRepository.cs
@@ -1,6 +1,9 @@
 ﻿using Hx.Abp.Attachment.Domain;
 using Hx.Abp.Attachment.Domain.Shared;
 using Microsoft.EntityFrameworkCore;
+using System.Linq;
+using System.Linq.Dynamic.Core;
+using Volo.Abp;
 using Volo.Abp.Domain.Repositories.EntityFrameworkCore;
 using Volo.Abp.EntityFrameworkCore;
 
@@ -49,30 +52,33 @@ namespace Hx.Abp.Attachment.EntityFrameworkCore
             bool includeDetails = true,
             CancellationToken cancellationToken = default)
         {
-            var inputReferences = inputs.Select(i => i.Reference).Distinct().ToList();
-            var inputReferenceTypes = inputs.Select(i => i.ReferenceType).Distinct().ToList();
-            return await (await GetDbSetAsync())
-                .IncludeDetails(includeDetails)
-                .Where(p => p.ParentId == null && inputReferences.Contains(p.Reference))
-                .Where(p => inputReferenceTypes.Contains(p.ReferenceType))
-                .OrderBy(d => d.Reference)
-                .ThenBy(d => d.SequenceNumber)
-                .ToListAsync(cancellationToken);
+            var queryable = (await GetDbSetAsync()).IncludeDetails(includeDetails);
+            var predicate = GetCataloguesByReference(inputs);
+            queryable = queryable.Where(predicate);
+            return await queryable.OrderBy(d => d.Reference)
+             .ThenBy(d => d.SequenceNumber)
+             .ToListAsync(cancellationToken);
         }
         public async Task DeleteByReferenceAsync(
             List<GetAttachListInput> inputs,
-            CatalogueCreateMode createMode,
             bool includeDetails = true,
             CancellationToken cancellationToken = default)
         {
-            var inputReferences = inputs.Select(i => i.Reference).Distinct().ToList();
-            var inputReferenceTypes = inputs.Select(i => i.ReferenceType).Distinct().ToList();
-            var ids = await (await GetDbSetAsync())
-                .IncludeDetails(includeDetails)
-                .Where(p => inputReferences.Contains(p.Reference))
-                .WhereIf(createMode == CatalogueCreateMode.Overlap, p => inputReferenceTypes.Contains(p.ReferenceType))
-                .Select(d => d.Id)
-                .ToListAsync(cancellationToken);
+            var queryable = (await GetDbSetAsync()).IncludeDetails(includeDetails);
+            var predicate = PredicateBuilder.New<AttachCatalogue>(true);
+            foreach (var input in inputs)
+            {
+                predicate = predicate?.Or(d =>
+                d.Reference == input.Reference &&
+                d.ReferenceType == input.ReferenceType);
+            }
+            if (predicate == null)
+            {
+                throw new UserFriendlyException("输入条件不能为空！");
+            }
+            queryable = queryable.Where(predicate);
+            var ids = await queryable.Select(d => d.Id)
+            .ToListAsync(cancellationToken);
             await DeleteManyAsync(ids);
         }
         public async Task<List<AttachCatalogue>> VerifyUploadAsync(
@@ -80,15 +86,12 @@ namespace Hx.Abp.Attachment.EntityFrameworkCore
             bool includeDetails = true,
             CancellationToken cancellationToken = default)
         {
-            var inputReferences = inputs.Select(i => i.Reference).Distinct().ToList();
-            var inputReferenceTypes = inputs.Select(i => i.ReferenceType).Distinct().ToList();
-            return await (await GetDbSetAsync())
-                .IncludeDetails(includeDetails)
-                .Where(p => p.ParentId == null && inputReferences.Contains(p.Reference))
-                .Where(p => inputReferenceTypes.Contains(p.ReferenceType))
-                .OrderBy(d => d.Reference)
-                .ThenBy(d => d.SequenceNumber)
-                .ToListAsync(cancellationToken);
+            var queryable = (await GetDbSetAsync()).IncludeDetails(includeDetails);
+            var predicate = GetCataloguesByReference(inputs);
+            queryable = queryable.Where(predicate);
+            return await queryable.OrderBy(d => d.Reference)
+             .ThenBy(d => d.SequenceNumber)
+             .ToListAsync(cancellationToken);
         }
         public override async Task<IQueryable<AttachCatalogue>> WithDetailsAsync()
         {
@@ -96,9 +99,47 @@ namespace Hx.Abp.Attachment.EntityFrameworkCore
             return queryable.IncludeDetails();
         }
 
-        public async Task<bool> AnyByNameAsync(string catalogueName, string reference)
+        public async Task<bool> AnyByNameAsync(string catalogueName, string reference, int referenceType)
         {
-            return await (await GetDbSetAsync()).AnyAsync(p => p.CatalogueName == catalogueName && p.Reference == reference);
+            return await (await GetDbSetAsync()).AnyAsync(p =>
+            p.CatalogueName == catalogueName &&
+            p.Reference == reference &&
+            p.ReferenceType == referenceType);
+        }
+        public async Task<List<AttachCatalogue>> AnyByNameAsync(List<GetCatalogueInput> inputs)
+        {
+            var predicate = PredicateBuilder.New<AttachCatalogue>(true);
+            foreach (var input in inputs)
+            {
+                predicate = predicate?.Or(d =>
+                d.Reference == input.Reference &&
+                d.ReferenceType == input.ReferenceType &&
+                d.CatalogueName == input.CatalogueName);
+            }
+            if (predicate == null)
+            {
+                throw new UserFriendlyException("输入条件不能为空！");
+            }
+            return await (await GetDbSetAsync()).Where(predicate).ToListAsync();
+        }
+        public async Task DeleteRootCatalogueAsync(List<GetCatalogueInput> inputs)
+        {
+            var queryable = (await GetDbSetAsync()).AsQueryable();
+            var predicate = PredicateBuilder.New<AttachCatalogue>(true);
+            foreach (var input in inputs)
+            {
+                predicate = predicate?.Or(d =>
+                d.Reference == input.Reference &&
+                d.ReferenceType == input.ReferenceType &&
+                d.CatalogueName == input.CatalogueName);
+            }
+            if (predicate == null)
+            {
+                throw new UserFriendlyException("输入条件不能为空！");
+            }
+            queryable = queryable.Where(predicate);
+            var ids = await queryable.Select(d => d.Id).ToListAsync();
+            await DeleteManyAsync(ids);
         }
         public async Task<int> GetMaxSequenceNumberByReferenceAsync(string reference)
         {
@@ -106,6 +147,22 @@ namespace Hx.Abp.Attachment.EntityFrameworkCore
                 .Where(d => d.CatalogueName == reference)
                 .OrderByDescending(d => d.SequenceNumber)
                 .Select(d => d.SequenceNumber).FirstOrDefaultAsync();
+        }
+        private ExpressionStarter<AttachCatalogue> GetCataloguesByReference(List<GetAttachListInput> inputs)
+        {
+            var predicate = PredicateBuilder.New<AttachCatalogue>(true);
+            foreach (var input in inputs)
+            {
+                predicate = predicate?.Or(d =>
+                d.ParentId == null &&
+                d.Reference == input.Reference &&
+                d.ReferenceType == input.ReferenceType);
+            }
+            if (predicate == null)
+            {
+                throw new UserFriendlyException("输入条件不能为空！");
+            }
+            return predicate;
         }
     }
 }
